nostr-relay: handle more ws messages and stop leaking connections
If a client disenganged or stopped working properly, we kept running
a task in tokio to handle the connection. This was keeping the server end
of the websocket open, along with multiple subscriptions to the
broadcast channel, thus leaking resources.

We now handle ws errors and Close message by stopping the tokio task,
thus dropping the ws connection and the internal brodcast subscribers.

We also actively send Ping requests that will cause the same kind of termination
if not answered within a specific timeframe.