nostr-relay: stop terminating websockets for valid errors
When handling EVENT requests, any kind of error would terminate the
websocket. For example, if for any (valid) reason an EVENT request
could not be accepted (e.g. invalid signature), instead of terminating
the websocket, we just reply with "OK" and specify the request has failed.
Before, we replied with OK only when it was accepted.

Also, any error with the broadcasting channel, would cause the websocket
to be terminated. This is fixed.