Client: merge 2 tasks in one select to simplify ws management
We need to be able to manage websocket in case of disconnections,
so it's cleaner to have both ends in the same task.

This means the internal channel must be sync (unbounded), so we
do not wait for send in one of the two select branches, while the
other one is supposed to be reading: this could result in a deadlock
given writers and readers are in the same selecy statement.