Assigner: be more tolerant when publish fails.
Assigner should not panic when it cannot publish, it could be a delay
in the ACK response or just server down and client reconnecting.

Decreased the tolerance for ACK in Client from 20 to 10s because
everything slows down when ACKs do not arrive promptly.
The logic in the client for ACKs must be changed because it's a real
bottleneck.